第二信号(无mv)传入:input;

+++ SMG_NothingIsAll/AIFoundation/AIAlg/AlgsModel/CustomAlgsModel/CustomAlgsChargeModel.h
@@ -10,6 +10,6 @@
 
 @interface CustomAlgsChargeModel : NSObject
 
-@property (assign,nonatomic) CGFloat value;
+@property (assign,nonatomic) float value;
 
 @end
